+++ hyperSpec/NEWS.md
@@ -13,6 +13,7 @@
 * Portions of `hyperSpec` were spun out into their own packages for ease of maintenance.  ELABORATE
 * Dataset `faux_cell` and function `generate_faux_cell` replace `chondro` dataset (#125, #156, #180, #229).
 * Dataset `chondro` was moved to package **hySpc.chondro** (https://r-hyperspec.github.io/hySpc.chondro/).
+* Function `spc.NA.linapprox()`, which was deprecated for long time, is now completely removed (#239).
 
 
 ## Non-User-Facing Changes from 0.99 Series